+++ trixnity-clientserverapi/trixnity-clientserverapi-server/src/commonMain/kotlin/net/folivo/trixnity/clientserverapi/server/MediaApiHandler.kt
@@ -42,6 +42,11 @@ interface MediaApiHandler {
     suspend fun downloadMedia(context: MatrixEndpointContext<DownloadMedia, Unit, Media>): Media
 
     /**
+     * @see [DownloadMediaWithFileName]
+     */
+    suspend fun downloadMediaWithFileName(context: MatrixEndpointContext<DownloadMediaWithFileName, Unit, Media>): Media
+
+    /**
      * @see [DownloadThumbnailLegacy]
      */
     @Deprecated("use downloadThumbnail instead")